member-quick-actions: move MemberAction + GetMemberActionsUseCase into shared module
